offroad-terrain-segmentation/requirements.txt
# Offroad Terrain Segmentation - Python Dependencies

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=9.0.0
numpy>=1.21.0

# Web API
flask>=2.3.0
flask-cors>=4.0.0

# Optional: For production deployment
# flask-limiter>=3.3.0        # Rate limiting
# flask-httpauth>=4.8.0       # Authentication
# gunicorn>=20.1.0            # Production WSGI server

# Optional: For enhanced functionality
# opencv-python>=4.7.0        # Advanced image processing
# matplotlib>=3.5.0           # Plotting and visualization
# scikit-image>=0.19.0        # Additional image utilities
# tqdm>=4.65.0                # Progress bars for batch processing
